Do not expose secret key (#3613)
